feat udistrital/acta_recibido_crud#60, udistrital/acta_recibido_crud#61 Se unifican componentes registro-acta y acta-especial
