--- src/app/pages/acta-recibido/acta-especial/acta-especial.component.scss
@@ -1,69 +1,0 @@
-:host /deep/ ng2-st-tbody-edit-delete {
-    display: flex !important;
-    height: 0 !important;
-  }
-
-  :host /deep/ ng2-st-tbody-custom a.ng2-smart-action.ng2-smart-action-custom-custom {
-    display: inline-block;
-    width: 30px;
-    text-align: center;
-    font-size: 1.1em;
-  }
-
-  :host /deep/ ng2-st-tbody-custom a.ng2-smart-action.ng2-smart-action-custom-custom:hover {
-    color: #5dcfe3;
-  }
-  :host /deep/ .material-icons {
-    font-family: 'Material Icons' !important;
-  }
-  .example-container {
-    height: 400px;
-    overflow: auto;
-  }
-  .bla {
-    position: relative;
-    z-index: 1;
-  }
-  .cell {
-    height: 50px;
-    width: 150px;
-    text-align: right;
-  }
-  .cell-2 {
-    height: 50px;
-    width: 150px;
-    text-align: left;
-  }
-  .cell-3 {
-    height: 40px;
-    width: 150px;
-    text-align: right;
-  }
-  table {
-    width: 100%;
-    text-align: center;
-  }
-  .campo {
-    color:#a61e1c;
-    text-decoration-color: #a61e1c;
-    text-decoration-line: underline;
-  }
-  .soporte {
-    float: none;
-    text-align: center;
-  }
-
-// Para que la lista de Dependencias no salga "recortada" (Ver Issue #315)
-:host /deep/ ng2-completer[formcontrolname="Dependencia"] .completer-dropdown-holder {
-  position: fixed;
-}
-:host /deep/ ng2-completer[formcontrolname="Dependencia"] .completer-dropdown {
-  max-height: 200px !important;
-  overflow-y: scroll;
-  overflow-x: hidden;
-}
-
-table.consolidado {
-  width: auto;
-  margin: auto;
-}
