--- src/app/pages/acta-recibido/acta-especial/acta-especial.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ActaEspecialComponent } from './acta-especial.component';
-
-describe('ActaEspecialComponent', () => {
-  let component: ActaEspecialComponent;
-  let fixture: ComponentFixture<ActaEspecialComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ActaEspecialComponent ],
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ActaEspecialComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
